Fixed situation for hover
Before it was possible to hover mouse on piece that had in beating range a piece that was checking his King but there were actually two pieces which were checking King, still it wasn't possible for that piece to move. It was only possible to choose it or hover mouse on it and it couldn't do anything.